# app/docs/swagger.yaml
definitions:
...
  handlers.BulkPutRequestEntry:
    properties:
      type:
        description: |-
          Optional document type (e.g., "json", "counter", "list", "set").
          If omitted, "json" is assumed.
        type: string
      value:
        description: |-
          The value to store for the document.
          Can be a string, number, object, array, etc.
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.createFamilyRequest:
    properties:
      name:
        example: logs
        type: string
    type: object
...
  model.Document:
    properties:
      key:
        description: Logical key
        type: string
      meta:
        allOf:
        - $ref: '#/definitions/model.Metadata'
        description: Metadata for versioning, expiration, etc.
      value:
        description: 'Content: string, int, []string, map[string]any, etc.'
    type: object
  model.Metadata:
    properties:
      created_at:
        description: When document was created
        type: string
      expiration:
        description: TTL as Unix timestamp (0 = never)
        type: integer
      rev:
        description: Revision ID for conflict resolution
        type: string
      type:
        description: Document type (json, counter, list, set)
        type: string
      updated_at:
        description: When document was last updated
        type: string
    type: object
info:
  contact: {}
paths:
...
  /documents:
    delete:
      description: Deletes a document by key within a specified column family.
      parameters:
      - description: Document key to delete
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Document successfully deleted
        "400":
          description: Missing or invalid parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document or column family not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete a document
      tags:
      - documents
    get:
      description: Retrieves a document by key, including its value and metadata.
      parameters:
      - description: Document key
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Optional RocksDB fill cache read option
        in: query
        name: fill_cache
        type: boolean
      - description: Optional RocksDB read tier (e.g. 'all', 'cache-only')
        in: query
        name: read_tier
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Document'
        "400":
          description: Missing or invalid key
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Retrieve a document
      tags:
      - documents
    post:
      consumes:
      - application/json
      description: Stores a document in the specified column family. Supports optimistic
        concurrency via CAS.
      parameters:
      - description: Document key
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Document type (e.g. 'json', 'counter', 'list')
        in: query
        name: type
        type: string
      - description: CAS (revision) for concurrency control
        in: query
        name: cas
        type: string
      - description: Document value (JSON-encoded)
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Document'
        "400":
          description: Invalid input or missing value
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Column family not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: CAS mismatch
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Store or update a document
      tags:
      - documents
...
  /documents/bulk/put:
    post:
      consumes:
      - application/json
      description: Stores multiple documents in a single call. Each document is defined
        by a key and a value/type pair.
      parameters:
      - description: Column family (defaults to 'default')
        in: query
        name: cf
        type: string
      - description: Map of key to value/type entry
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            $ref: '#/definitions/handlers.BulkPutRequestEntry'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/model.Document'
            type: object
        "400":
          description: Invalid input or empty payload
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Bulk insert documents
      tags:
      - documents
...
  /documents/lists/pop:
    post:
      consumes:
      - application/json
      description: Removes and returns the last element of a list-type document.
      parameters:
      - description: Key of the list document
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: 'Write option: sync write to disk'
        in: query
        name: sync
        type: boolean
      - description: 'Write option: disable write-ahead log'
        in: query
        name: disable_wal
        type: boolean
      - description: 'Write option: disable slowdown on write buffer full'
        in: query
        name: no_slowdown
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Returns the popped element
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or missing key
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Pop element from list
      tags:
      - lists
...
  /documents/replace:
    post:
      consumes:
      - application/json
      description: Replaces a document if it already exists. Fails if the key does
        not exist. Supports CAS for concurrency control.
      parameters:
      - description: Document key
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Document type (e.g. 'json', 'counter', 'list')
        in: query
        name: type
        type: string
      - description: CAS (revision) for concurrency control
        in: query
        name: cas
        type: string
      - description: New value for the document
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Document'
        "400":
          description: Invalid request or missing value
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Key not found or column family missing
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: CAS mismatch
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Replace an existing document
      tags:
      - documents
...
  /families:
    get:
      description: Retrieves the names of all available column families.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List column families
      tags:
      - families
    post:
      consumes:
      - application/json
      description: Creates a new column family with the specified name.
      parameters:
      - description: Name of the column family to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createFamilyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created column family name
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid or missing column family name
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Column family already exists
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create column family
      tags:
      - families
...
swagger: "2.0"